digests unit tests into sub packages
